requirements_build.txt
accelerate
bitsandbytes
datasets
einops
hydra-core
jinja2
kaggle
loguru
omegaconf
python-dotenv
transformers
peft==0.9.0
wandb